validações

+++ bank/accounts/models.py
@@ -1,9 +1,11 @@
 from django.db import models
 
 # Create your models here.
+from rest_framework import serializers
 
 
 class Account(models.Model):
     owner = models.CharField(max_length=200, blank=True, default='')
-    balance = models.DecimalField(decimal_places=2, max_digits=10)
+    balance = models.FloatField()
     creation_date = models.DateTimeField(auto_now_add=True)
+
